refactor(repl): move signal handling in command
